Add option to migrate entire db or only update home_user

+++ db_setup.py
@@ -3,26 +3,45 @@
 # * Export your db to have a backup before running.
 
 import os
+import sys
 import web
+from django.core.wsgi import get_wsgi_application
 from planetterp.config import USER, PASSWORD
 
-print("Wiping django tables...")
-os.system("python3 wipe_django.py")
+# https://stackoverflow.com/a/43391786
+os.environ['DJANGO_SETTINGS_MODULE'] = 'planetterp.settings'
+application = get_wsgi_application()
 
-print("Updating web.py tables...")
-db_name = os.environ.get("PLANETTERP_MYSQL_DB_NAME", "planetterp")
-db = web.database(dbn='mysql', db=db_name, user=USER, pw=PASSWORD, charset='utf8mb4')
+from home.models import User
 
-# Make the changes mentioned in webpy_to_django.py
-db.query('DELETE FROM planetterp.professors WHERE id < 0')
-db.query('UPDATE planetterp.professors SET verified = 1 WHERE verified IS NULL')
-db.query('UPDATE planetterp.reviews SET verified = 1 WHERE verified IS NULL')
-db.query("UPDATE planetterp.users SET email = '' WHERE CHARACTER_LENGTH(email) > 254")
+if "full" in sys.argv:
+    print("Starting full setup...\nWiping django tables...")
+    os.system("python3 wipe_django.py")
 
 
+    print("Updating web.py tables...")
+    db_name = os.environ.get("PLANETTERP_MYSQL_DB_NAME", "planetterp")
+    db = web.database(dbn='mysql', db=db_name, user=USER, pw=PASSWORD, charset='utf8mb4')
+
+    # Make the changes mentioned in webpy_to_django.py
+    db.query('DELETE FROM planetterp.professor_courses WHERE professor_id < 0')
+    db.query('DELETE FROM planetterp.reviews WHERE professor_id < 0')
+    db.query('DELETE FROM planetterp.grades WHERE professor_id < 0')
+    db.query('DELETE FROM planetterp.grades_historical WHERE professor_id < 0')
+    db.query('DELETE FROM planetterp.professors WHERE id < 0')
+
+    db.query('UPDATE planetterp.professors SET verified = 1 WHERE verified IS NULL')
+    db.query('UPDATE planetterp.reviews SET verified = 1 WHERE verified IS NULL')
+    db.query("UPDATE planetterp.users SET email = NULL WHERE CHARACTER_LENGTH(email) > 254 OR email = ''")
 
 
-print("Migrating to django. This might take a few minutes...")
-os.system("python3 webpy_to_django.py")
+    print("Migrating to django. This might take a few minutes...")
+    os.system("python3 webpy_to_django.py")
+else:
+    print("Starting django setup...\nUpdating user emails...")
+    User.objects.filter(email='').update(email=None)
+    os.system("python3 manage.py makemigrations")
+    os.system("python3 manage.py migrate")
+
 
 print("Done")
